move stuff around inside test-common
