+++ test-common/canton/it-lib/src/main/com/daml/TestTemplateIdentifiers.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.testing
+package com.daml
 
 import com.daml.ledger.api.v1.value.Identifier
